Add: PolygonCalculator.doubleToLong, .doubleToLongTest, setDeltaXY, setDeltaXYTest  setCorrectionXY, setCorrectionTestResiduals

+++ geocalculator_service/src/main/java/ru/taheoport/geocalculator_service/model/Residuals.java
@@ -37,7 +37,7 @@ public class Residuals {
     /**
      * Actual absolute linear mistake
      */
-    private double absolute;
+    private long absolute;
 
     /**
      * Actual relative linear mistake
